Move fetchProfile actions to user and update reducer
