+++ src/modules/dashboard/dashboardActions.js
@@ -2,25 +2,6 @@ import * as types from './dashboardTypes';
 
 import { axiosWithAuth } from '../../utils/auth';
 
-export const fetchProfile = () => dispatch => {
-  dispatch({ type: types.ON_BEGIN_PROFILE_FETCH });
-
-  axiosWithAuth()
-    .get(`/auth/view_profile`)
-    .then(({ data }) => {
-      dispatch({
-        type: types.ON_PROFILE_FETCH_SUCCESS,
-        payload: data.user,
-      });
-    })
-    .catch(err => {
-      dispatch({
-        type: types.ON_PROFILE_FETCH_FAILED,
-        payload: err,
-      });
-    });
-};
-
 export const fetchTags = () => dispatch => {
   dispatch({ type: types.ON_START_FETCHING_TAGS });
