+++ src/modules/dashboard/dashboardReducer.js
@@ -1,7 +1,6 @@
 import * as types from './dashboardTypes';
 
 const initialState = {
-  userProfile: {},
   loading: false,
   errors: null,
   creatingDeck: false,
@@ -14,26 +13,6 @@ const initialState = {
 
 const dashboardReducer = (state = initialState, action) => {
   switch (action.type) {
-    case types.ON_BEGIN_PROFILE_FETCH:
-      return {
-        ...state,
-        loading: true,
-      };
-
-    case types.ON_PROFILE_FETCH_FAILED:
-      return {
-        ...state,
-        loading: false,
-        errors: action.payload,
-      };
-
-    case types.ON_PROFILE_FETCH_SUCCESS:
-      return {
-        ...state,
-        loading: true,
-        userProfile: action.payload,
-      };
-
     case types.ON_START_FETCHING_TAGS:
       return { ...state, fetchingTags: true, loading: true };
